Bugfixing and Tidying up
