+++ ApsimNG/Views/ManagerView.cs
@@ -5,10 +5,8 @@ using UserInterface.Interfaces;
 
 namespace UserInterface.Views
 {
-    
     public class ManagerView : ViewBase,  IManagerView
     {
-
         private PropertyView propertyEditor;
         private IEditorView scriptEditor;
         private Notebook notebook;
@@ -29,10 +27,9 @@ namespace UserInterface.Views
             propertyEditor = new PropertyView(this);
             scriptEditor = new EditorView(this)
             {
-
                 ShowLineNumbers = true,
                 Language = "c-sharp",
-
+                ReadOnly = true
             };
             notebook.AppendPage(propertyEditor.MainWidget, new Label("Parameters"));
             notebook.AppendPage(((ViewBase)scriptEditor).MainWidget, new Label("Script"));
@@ -82,6 +79,7 @@ namespace UserInterface.Views
             } 
             else
             {
+                scriptEditor.ReadOnly = false;
                 notebook.Drawn -= OnDrawn;
                 if (cursor != null && this.TabIndex == TAB_SCRIPT)
                     scriptEditor.Location = cursor;
@@ -135,31 +133,5 @@ namespace UserInterface.Views
 
         public IPropertyView PropertyEditor { get { return propertyEditor; } }
         public IEditorView Editor { get { return scriptEditor; } }
-    }
-
-    public interface IManagerView
-    {
-        /// <summary>
-        /// Provides access to the properties grid.
-        /// </summary>
-        /// <remarks>
-        /// Change type to IProeprtyView when ready to release new property view.
-        /// </remarks>
-        IPropertyView PropertyEditor { get; }
-
-        /// <summary>
-        /// Provides access to the editor.
-        /// </summary>
-        IEditorView Editor { get; }
-
-        /// <summary>
-        /// Indicates the index of the currently active tab
-        /// </summary>
-        int TabIndex { get; set; }
-
-        /// <summary>
-        /// The values for the cursor and scrollbar position in the script editor
-        /// </summary>
-        ManagerCursorLocation CursorLocation { get; set; }
     }
 }
